changing the config system for the tutorial
we moved from the enum and switch/case based system to one
that only uses strings. This seems a lot easier, and in particular,
it saves the user from having to mess with yet another file where
stuff needs to be configured.

Also, the enum based system is very tricky to break down to a
modular level- so far, all enums had to be globally defined for
all combination projects...