Reactivating 2D plugin scans
One can now do 2D plugin scans in the same way as 1D ones.

Improved the 2D plotting class to cope with 2D plugin plots.

Changed and homogenized the code in GammaComboEngine that
defines the work flow steered by the command line arguments
-a plugin, -a pluginbatch, etc. For the user this should
be transparent.

Change a lot of indents.