perhaps fixing the chi2 explosion
When the scan encounteres a failed fit that typically gives chi2=-1e30,
it used to set it to 0, because thats the lowest allowed value. But this
is also corresponds to p-value=1, screwing up the plots. Now we set it
to the best fit chi2 found so far, plus 25 units, corresponding to 5
sigma.

Changed scaleDownErrors() a bit, but this doesn't affect anything.