adding a function to get a sub correlation matrix
adding getErrorSourceString()
adding getCorrelationSourceString()
adding getSubCorrelationStat()
adding getSubCorrelationSyst()
adding getSubMatrix()

indentation

changing 'firstScan' variable to 'firstScanDone' which is a clearer name

changing the z axis range of 2D chi2 control plots from 36 units to 25
which corresponds to 5 sigma

changing the z axis range of 2D chi2 full plots from 36 units to 25
which corresponds to 5 sigma

adding getSubCorrelationStat()
adding getSubCorrelationSyst()
adding getSubMatrix()

These functions take a list of row indices and return a sub matrix
of the respective correlation matrix. This is useful to easily create
a new PDF class, that corresponds only to a subset of the observables,
without having to inherit from the large one (bad style). An example
is the DKDpi PDF in the gamma combination, and the DK version of it,
that only contains the DK observables and discards the Dpi ones.