fixed a problem when using --scanrangey in default 2D scans
The problem was that one got a warning when using --scanrangey to change
the y scan range from the inital 1D scan, that are part of the default
2D scan strategy. The 1D scanner for the y variable didn't get the
--scanrangey arguments propagated, causing the warning that the range
was changed after initScan() was called, and also a chi2 explosion for
the combination I tested this with.

Fixed by introducing 2 new functions that set the range, and remember
the state- so initScan() doesn't reset the range if it was set manually.

new functions in MethodAbsScan:

setXscanRange(float min, float max)
setYscanRange(float min, float max)

Also improving some printout messages.