Fixing a bug with 2D scan ranges, deprecating old constructors
Fixing a bug with the scan ranges: they got set to the default values
defined in OptParser, because the if statements where missing from the
MethodAbsScan::setXscanRange() and MethodAbsScan::setYscanRange()
function, that switch them to transparency when min==max.

To help debugging this, the code was changed such that
MethodAbsScan::initScan() can only be called once. If called twice,
there will be an error and exit. Also, the default values in OptParser
now have different bogus values to help tracking problems back.

Old constructors of MethodAbsScan were depricated (exit), which don't
initialize all members. So far the code still runs so maybe we don't
need them anymore.