adding commented code for future use
The code is aimed at improving the --magnetic behaviour.
So far, without much success.

Added a version of ConfidenceContours::addBoundaryBins() that not only
adds one boundary bin so that the contours close, but two. The innermost
of which is filled with a copy of the outer bins of the original
histogram. The hope was that then the contours close far enough outside
the nominal plotted area so that the --magnitc magic was not needed
anymore. Doesn't it work 100%.

The other two commented functions implement the solutions for a straight
line extrapolation based on two points.