Fixing a bug: plots where drawn in wrong canvasses
After the redesign of the plot mechanism, i.e. that the plots now have
their main canvas not created in the constructor, a few more cd()
statements where needed.

In particular: when running on multiple combiners and requesting pull
plots (-c 1 -c2 --pulls) for 1D, the 1-CL curves ended up in the pull
canvas...