fixing a problem where the wrong pol2 solution was picked
There was still some logic in the code that tried to pick the correct
solution based on the central value. This failed today. The easier
solution should be to just pick the solution that is closest to the bin
we interpolated around. This was a fallback already present in the code,
and is now the default.