fixing a bug where solutions were drawn at (0,0)
There was a bug in the code that was adding the solution to the graph of
the 1-CL curve, in order to always have the solution exactly at 1.0. It
failed when there was no solution, or when the solution was at the plot
boundary.

The new code now lives in its own function. Much better.