Need to change cov matrix build to virtual func
There are actually cases where the user might want to overwrite this
function as is done the e.g. building the PDF. An example is when
making assumptions / studies with zero systematics and very many numbers
of observables (i.e. 2D binned distribution). In this case the default
cov matrix inversion is killed by the accuracy falling below threshold.